feat(portal): derive tier from wallet and drop the link layer on SaaS
Removes the account-link concept from the SaaS portal so it isn't just
mounted-but-inert: the "Link Stirling account" CTA is gone, the tier
reflects the real subscription, and the link machinery tree-shakes out of
the SaaS bundle. Self-hosted is unchanged. Four file-override seams, no
runtime flavor flags.

- usePlanTier seam: TierContext no longer reads LinkContext. Self-hosted
  derives the tier from the link/subscription state; SaaS derives it from
  the wallet (subscribed -> pro, else free). Fixes the sidebar showing the
  free "Docs processed" meter for subscribed SaaS orgs.
- PortalProviders seam: self-hosted mounts the account-link layer
  (LinkProvider / AccountLinkProvider / login modal) around the shared
  PortalChrome; SaaS mounts TierProvider -> UIProvider -> PortalChrome
  only. PortalApp slims to Theme/Mantine/AuthBoundary + <PortalProviders>.
- LinkAccountFooterItem seam: the sidebar CTA; SaaS override renders null.
- accountLinkSettings seam: the Settings "Account link" section; SaaS
  override is null, so the nav item and its panel both drop and the
  link-only AccountLinkPanel is never imported into the SaaS bundle.

Verified: typecheck all 7 variants, vitest portal+saas (140), saas +
proprietary builds, eslint + prettier — all green.
