+++ frontend/editor/src/saas/portal/components/LinkAccountFooterItem.test.tsx
@@ -1,0 +1,10 @@
+import { describe, expect, it } from "vitest";
+import { render } from "@testing-library/react";
+import { LinkAccountFooterItem } from "@portal/components/LinkAccountFooterItem";
+
+describe("LinkAccountFooterItem (SaaS)", () => {
+  it("renders nothing — SaaS has no account to link", () => {
+    const { container } = render(<LinkAccountFooterItem />);
+    expect(container).toBeEmptyDOMElement();
+  });
+});
