+++ frontend/editor/src/saas/portal/components/LinkAccountFooterItem.tsx
@@ -1,0 +1,7 @@
+/**
+ * SaaS has no account-link concept — the signed-in account IS the SaaS account,
+ * so there is nothing to link and no footer CTA to show.
+ */
+export function LinkAccountFooterItem() {
+  return null;
+}
